requirements.txt
streamlit
pandas
matplotlib
biopython
requests